+++ tests/test_block_factory.py
@@ -310,3 +310,85 @@ def test_embedding_projection():
     input_mask = torch.randn(SEQ, dtype=torch.float, device=device)
     input_mask[input_mask < 0.0] = -float("inf")
     _ = block(inputs, input_mask=input_mask)
+
+
+@pytest.mark.parametrize("device", DEVICES)
+@pytest.mark.skipif(
+    not torch.cuda.is_available(), reason="This test requires a CUDA device"
+)
+def test_simplicial_embedding(
+    device: torch.device,
+):
+    attention_config = {
+        "name": "scaled_dot_product",
+        "dropout": 0.1,
+        "causal": False,
+        "window_size": SEQ // 8 + 1,
+        "seq_len": SEQ,
+        "dim_model": MODEL,
+        "num_heads": 4,
+    }
+
+    multi_head_config = {
+        "num_heads": 4,
+        "dim_model": MODEL,
+        "residual_dropout": 0.1,
+        "attention": attention_config,
+    }
+
+    feedforward_config = {
+        "name": "MLP",
+        "dim_model": MODEL,
+        "dropout": DROPOUT,
+        "activation": "relu",
+        "hidden_layer_multiplier": 4,
+    }
+
+    position_encoding_config = {
+        "name": "sine",
+        "dim_model": MODEL,
+        "seq_len": SEQ,
+        "vocab_size": VOCAB_SIZE,
+    }
+
+    block_config = xFormerEncoderConfig(
+        dim_model=MODEL,
+        multi_head_config=multi_head_config,
+        feedforward_config=feedforward_config,
+        position_encoding_config=position_encoding_config,
+        layer_norm_style="pre",
+        reversible=False,
+        simplicial_embeddings={"L": 4},
+    )
+
+    # Test that the whole block can be instantiated
+    block = xFormerEncoderBlock.from_config(block_config).to(device)
+
+    # Check that the dimensions make sense, to a FW pass
+    inputs = torch.rand(BATCH, SEQ, device=device)
+    _ = block(inputs)
+
+    # Check that we support attention masking, at least interface wise (do not check correctness yet)
+    att_mask = torch.ones(SEQ, SEQ, dtype=torch.bool, device=device)
+    _ = block(inputs, att_mask=att_mask)
+
+    # Check that we support input masking, at least interface wise (do not check correctness yet)
+    input_mask = torch.randn(SEQ, dtype=torch.float, device=device)
+    input_mask[input_mask < 0.0] = -float("inf")
+    _ = block(inputs, input_mask=input_mask)
+
+    # Check that a faulty L is caught
+    block_config = xFormerEncoderConfig(
+        dim_model=MODEL,
+        multi_head_config=multi_head_config,
+        feedforward_config=feedforward_config,
+        position_encoding_config=position_encoding_config,
+        layer_norm_style="pre",
+        reversible=False,
+        simplicial_embeddings={"L": 3},
+    )
+
+    # Test that the whole block can be instantiated
+    with pytest.raises(AssertionError):
+        block = xFormerEncoderBlock.from_config(block_config).to(device)
+        _ = block(inputs)